Move home resources fetch in its own hook
